pyproject.toml
[tool.poetry]
name = "todolist"
version = "0.2.0"
description = "A task management CLI application with PostgreSQL backend"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "todolist", from = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
# Database dependencies
sqlalchemy = "^2.0.23"
psycopg2-binary = "^2.9.9"
alembic = "^1.13.0"
# Utilities
python-dotenv = "^1.0.0"
schedule = "^1.2.0"
fastapi = "^0.122.0"
uvicorn = {extras = ["standard"], version = "^0.38.0"}

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-cov = "^4.1.0"
black = "^23.11.0"
flake8 = "^6.1.0"
mypy = "^1.7.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
todolist = "todolist.main:main"
